+++ src/Entity/Topic.php
@@ -25,7 +25,7 @@ class Topic
     #[ORM\Column(length: 255, nullable: true)]
     private ?string $image = null;
 
-    #[ORM\ManyToOne(inversedBy: 'topics')]
+    #[ORM\ManyToOne(inversedBy: 'topic')]
     private ?User $user = null;
 
     #[ORM\OneToMany(mappedBy: 'topic', targetEntity: Reply::class)]